Enhance expansion and simplification of Ops
Defines some new operations for Algebra, changes the order of tuples
in the list representation of a sum of products of operators, and
adds cleanup functions to get to a reasonable simplification of Ops.
Still doesn't put scalars in canonical form, and hasn't provided any
expansion or simplification for the tensor product of operators.