Rename and tweak some functions on Op and OpAB
Renames bindScalars functions to be more explicit regarding their
domains and modify their action on scalar multiplication so they
aren't thwarted by nested scalar multiplies.

Also modify action of listDistributeOp(AB) on scalar multiplication
so that bindScalars may not be necessary, and use liftM2 (,) in
place of sequence to generate the Cartesian product for a tensor
product of sums.